[wasmfx] Fix allocation during stack switch
Tagged values in the stack-allocated argument buffer are not tracked by
the GC, so there must not be any GC until the buffer is unpacked by the
target stack.
In the suspend instruction, the new continuation object was mistakenly
allocated after initializing the argument buffer which could cause a GC.
Move it up.

R=clemensb@chromium.org

Fixed: 466948626
Change-Id: Ic97320d1c70d196b78f85e2aa1806a2dade8dd2c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7241572
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104198}