[iterator-helpers] Fix spec deviations in return method
This CL fixes the implementation of the `return` method for simple
iterator helpers (like those corresponding to .map, .filter, etc.)
and the specialized `flatMap` helper, so as to correctly mark the
helper's state as "executing" before calling an underlying iterator's
`.return()` method. It also makes sure that this is done correctly
if the underlying iterator is inactive (i.e. in SUSPENDED-START).
To do so, it adds what is necessary for distinguishing between
SUSPENDED-START and SUSPENDED-YIELD.

Moreover, it fixes two cleanup bugs in the `flatMap` helper, ensuring
that if closing an inner iterator throws an exception, the outer
iterator is properly closed and marked as exhausted and, if multiple
exceptions occur in this process, the correct one is thrown.

Bug: 463528051
Change-Id: I3d68c60b12eac336853823af94ccbf51a11efe4e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7223551
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104228}