[js_fuzzer] Enable using Babel transpiler when loading sources
This adds occasional usages of the ES6->ES5 transpiler for classes
when loading sources that emits a common pattern that might stress
optimizations of consecutive assignments to x.prototype.

We run the necessary Babel plugins with a probability of 10% on all
sources. Code in the wild typically is more minimized than the
typically Babel output. Therefore we also apply an inlining
minimization, necessary to trigger V8's optimization.

Bug: 429332174
Change-Id: I630ce1d9aaadc0610992378016b3c47e5648673e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7232292
Reviewed-by: Danylo Mocherniuk <mdanylo@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104222}