[tools] add a jstc command to debugger scripts
The jst command prints the stack traces in verbose mode, which can
be too verbose when printing a long JS stack whose arguments
include objects with a lot of properties. This adds a jstc command
to only print the stack trace in concise mode (without printing the
arguments) to complement it.

Change-Id: Ie3676122ae7f031182f09dd5f8c4b8bd02cf860c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7251009
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Joyee Cheung <joyee@igalia.com>
Cr-Commit-Position: refs/heads/main@{#104274}