[*lev] Fix LoadType-based phi untagging
It's possible for LoadTypeFields with non-smi type to be untagged if
they are the forward edge of a loop phi. We thus need to check for
this case first, because checking for the LoadTypeField case.

In a subsequent CL, we should generalize the loop phi optimization
that checks at the type of the inputs to all phis (and not just loop
phis), which will render obsolete the current special casing for
LoadTaggedField.

Fixed: 467844027
Change-Id: I5e258741b5065de15ac81eba01693163fa07e753
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7252209
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104299}