[api,execution] Print stack trace on OOM errors
We used to print the stack in the past already before crashing with
an OOM error until it was removed because it could trigger heap
allocations. This CL re-introduces this feature but avoids heap
allocations at the cost of potentially missing source locations. The
stack is now also reported as a crash key for convenience.

Bug: 389951203
Change-Id: I5f0b8c8a0124d3c9a987621b9b2b9c4df82d61c8
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7247091
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104312}