[runtime] Closure Creation from SetPrototypeProperties should be deferred to first call
This CL defers the creation of closures from `SetPrototypeProperties` to
when the function is first called.

When `SetPrototypeProperties` is used to assign multiple functions to a
prototype, instead of creating a `JSFunction` object for each one, this
change stores the `SharedFunctionInfo` directly on the object.

The `JSFunction` is only instantiated on the first access.This is a
performance optimization controlled by the new flag
`proto_assign_seq_lazy_func_opt`.

To support this lazy creation, a new `PrototypeSharedClosureInfo` struct
is introduced to store the necessary context for when the closure is
eventually created. The property access and IC mechanisms have been
updated to handle this deferred instantiation.

Bug: 449885256
Change-Id: Ia54b695940c0231e50f18e2509888c19f172968b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7026131
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Raphael Herouart <rherouart@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104322}