[heap] Remove pretenuring support in full GCs
In general the vast majority of GCs are Minor GCs, so not updating
AllocationSites during Major GCs shouldn't change pretenuring
heuristics much. Removing pretenuring support from Major GCs also
allows us to get rid of the Zombie state for AllocationSites. Without
the zombie state we also don't need to keep AllocationSites for
an additional GC cycle, so we do not need to mark AllocationSites
in ClearNonLiveReferences() anymore.

Removing the support for pretenuring in full GC doesn't break
resetting the pretenure decision at full GCs in
Heap::EvaluateOldSpaceLocalPretenuring. This is because this is
purely based on heap sizes before and after the full GC.

Bug: 463528049
Change-Id: Iaa673f6f0a24c9421c7477a2d503e9de3f7e5cd6
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7245796
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104330}