[heap] Make read-only MemoryChunk field conditionally optional
For the configuration with V8_CONTIGUOUS_COMPRESSED_RO_SPACE the RO
space is always allocated within the first
`kContiguousReadOnlyReservationSize` bytes.

This means that checking RO containment can be simplified to
1. A check against kContiguousReadOnlyReservationSize for cases that
   are guaranteed to rely on data cage or trusted cage decompression.
2. A check against kContiguousReadOnlySpaceMask for arbitrary full
   addresses, including InstructionStream objects in the CodeRange.

This makes the READ_ONLY_HEAP bit obsolete in this configuration.

The CL adjusts the callsites in generated code, rewires the C++
accessors, and finally puts the bit behind the compile-time flag.

Change-Id: Ia00fc9d27b8a68a77c4cc88ef002e9967acb4346
Bug: 429538831
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6905432
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104378}