[assembler] Fix code comments without source position
The `FileName()` of an empty (default initialized) `SourceLocation` is
the empty string, but not nullptr.
To check if a filename is set, we would need to check if the first
character is '\0', or we check the line number instead, which is what
`SourceLocation::ToString` already did internally.

For convenience (to make it easier to use the right check the next
time), this introduces an `operator bool` on `v8::SourceLocation` to
check if the source location is initialized.

R=leszeks@chromium.org

Change-Id: I87286600eaa26dbec57919ea83ecefd05368b0b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7264669
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104351}