[regexp][x64] Free up rcx as a scratch register.
The variable shift instructions all use rcx as a fixed
register, so it's inconvenient if that register is used
for the backtrack stack.  This change uses rbx as the
backtrack stack pointer, freeing up rcx as a scratch
register that can be used for shifting.

Change-Id: I932862da87211e17a8b95b3e6e6500abc102d8e1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7264344
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104349}