[wasm][revec] Fix invalid constant memory indices
Normally consecutive loads and stores will share identical index. It's
rare we need to have additional offset from the constant index.
To support it, we need to also update the offsets when reduce the loads
or stores, otherwise it will cause error. We may consider to support it
when we found it's really necessary.

Fixed: 467628265
Bug: 42202660
Change-Id: I1e7f311d9d92989777d8f0a6c19aed87e526646b
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7268684
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Chen, Yolanda <yolanda.chen@intel.com>
Reviewed-by: Xu, Hao A <hao.a.xu@intel.com>
Cr-Commit-Position: refs/heads/main@{#104388}