[turboshaft] Tune done --turboshaft_verify_reductions
This used to verify that output representations are preserved during
lowerings in CopyingPhase, but this doesn't seem like a great idea in
general: what if we want to optimize a Float64Add to a Int32Add? Or,
what if the Revectorizer replaces a Simd128 operation by a Simd256?

We already have verification that inputs have the correct
representation, which should catch almost all of the bugs that
--turboshaft_verify_reductions was catching. So, this CL tunes down
what this verification does: it now only verifies that output count
is preserved by CopyingPhase (which seems like a somewhat useful check,
since it caught a minor performance issue in
https://crbug.com/462045796).

Fixed: 468987574
Change-Id: I27d830000f1df81b6a7d729a34b0418100102f96
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7268805
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Darius Mercadier <dmercadier@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104369}