[codegen] Fix nondeterminism in non-avx code
In `I16x8DotI8x16I7x16S` for non-avx, `dst` and `src1` can not overlap
if `dst` does not equal `src2` (because the `movdqa` would overwrite
`src1` before it is being used).
This CL adds a missing DCHECK for that, and fixes this for Turbofan by
using `DefineSameAsInput(node, 1)` (completely avoiding the `movdqa`).
Liftoff just moves `lhs` (a.k.a. `src1`) to a scratch register if
necessary.

R=dmercadier@chromium.org

Fixed: 469286969
Change-Id: Ibdff833e852be7e37525637118b37888fbdc218d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7264670
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104370}