[wasm interpreter] Direct CallRef when the wasm instance is same.
To pass the return_call_ref case in wasm spec test, we need to tail
call the target function with a wasm func ref. This make us do a
direct call when the instance of target wasm function is the same as
current wasm instance.

Change-Id: I347bb8b97c79f3414d25eba393d1ded24afc28c1
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7247407
Reviewed-by: Paolo Severini <paolosev@microsoft.com>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Commit-Queue: Paolo Severini <paolosev@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#104432}