[sandbox] Handle SIGILL in the sandbox crash filter
Previously we didn't handle SIGILL as it might be caused by us executing
random invalid machine code, which is probably bad and should be treated
as a sandbox violation. However, there are also many cases of SIGILL
that are harmless, mostly because the compiler inserted a `ud2`
instruction to for example implement a release-mode CHECK.

As such, with this CL our crash filter now also handles SIGILL and
attempts to determine whether it was caused by a `ud2` or something
else by reading the faulting address (which will point to the faulting
instruction). In case of `ud2` we filter the crash, otherwise we report
a sandbox violation.

This CL also adds testcases for Fuzzilli to ensure that we can correctly
handle both cases.

Drive-By: Refactor CrashFilter to use a switch over the signal number
for better readability. Also let FilterCrash print the "Exiting process"
message instead of repeating it in every call to FilterCrash.

Bug: 42202821
Change-Id: Ibae32fb249dd3f1c4ac68dc4dc9d61fb5fe8ed6d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7269586
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104377}