[shared-strings] Ensure no stale forwarding indices in forwarding table
The StringForwardingTable can contain entries that are only kept alive
by other entries in the StringForwardingTable.
Depending on the order of iteration, we might not handle a string that
is not considered alive, but later kept alive by another entry.
In general this is not an issue, but we need to ensure that the hash for
these strings is a valid non-forwarding index hash.
Otherwise we could get strings with stale forwarding indices.

Fixed: 468231509
Change-Id: Iaae1d00959dc5f31ffbd9c89d5be79dc0d590f25
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7269784
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104376}