[fuzzilli] Fix build error on non-x64 machines
There we had a `break` after a FATAL, which caused -Wunreachable errors.
Instead, we now print a message since that will cause the crash tests to
fail, indicating that the crash tests need to be modified to support the
current architecture if ever needed.

Bug: 42202821
Change-Id: I3ddf90315326b3f2877e47b3020473796e3f0ef9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7273206
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Samuel Groß <saelo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104413}