Fix shared value barrier
Since http://crrev.com/c/6905432 the barrier would bail out if the
value is already in shared space. This is wrong as we do allocate
non-shared strings in shared space which need to be migrated for
sharing.

The CL reverts back to the initial state with also incorporating the
new way of checking for read-only.

Bug: 429538831
Change-Id: I9bea4c41f124d634d968d0e892a31326825e9568
Fixed: 469861341
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7274602
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104411}