[execution] Limit stack trace length on OOM crash
This CL adds a maximum length field to BuildMinimalStack(). This
is useful when building the stack trace for OOM crashes. The stack
trace is reported as a crash key in the end and we cannot report
more than 1024 bytes currently anyway.

In addition, this CL also uses kExposeFramesAcrossSecurityOrigins
when iterating the stack. This fixes crashes were isolate->context()
isn't initialized.

Bug: 389951203, 469644789
Change-Id: I3df045b9c110f24b3808549e20009706c46fdae7
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7277347
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104429}