[snapshot] Don't rewrite SFI IDs in read-only snapshot.
This is part of an effort to reduce the amount of
rewriting of objects in the read-only snapshot with the
eventual aim of sharing the read-only pages between
processes and between isolate groups. The hope is that
read-only pages can eventually be mapped directly from
the executable, which will make them shared.

Instead of writing a new shared function info ID into
each object when deserializing the read-only snapshot,
we just record the highest sfi in the read-only snapshot.
This is sufficient to ensure no IDs from the read-only
snapshot are reused later.

Change-Id: I637874bbe509556bf066fd1cd644588f0176c727

Bug: 472359966
Change-Id: I637874bbe509556bf066fd1cd644588f0176c727
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/7363773
Commit-Queue: Erik Corry <erikcorry@chromium.org>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#104475}